requirements/base.txt
pytz>=2019.3,<2020.1  # https://github.com/stub42/pytz
python-slugify>=4.0.0,<5.0.0  # https://github.com/un33k/python-slugify
Pillow>=7.0.0,<8.0.0  # https://github.com/python-pillow/Pillow
rcssmin>=1.0.6,<2.0.0  # https://github.com/ndparker/rcssmin
argon2-cffi>=19.2.0,<19.3.0  # https://github.com/hynek/argon2_cffi
whitenoise>=5.0.1,<6.0.0  # https://github.com/evansd/whitenoise
# redis>=3.4.1 # https://github.com/andymccurdy/redis-py

# Django
# ------------------------------------------------------------------------------
#django==2.1.1
#django==3.0.4 #,<3.1.0  # pyup: < 3.1  # https://www.djangoproject.com/
django>=3.0.4,<3.1.0  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ>=0.4.5,<0.5.0  # https://github.com/joke2k/django-environ
django-model-utils>=4.0.0,<5.0.0  # https://github.com/jazzband/django-model-utils
# django-allauth>=0.41.0,<0.42.0  # https://github.com/pennersr/django-allauth
git+https://github.com/nad2000/django-allauth.git@rapidconnect#egg=django_allauth
PyJWT>=1.7.1,<1.8.0
django-crispy-forms>=1.9.0,<2.0.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor>=2.4,<3.0  # https://github.com/django-compressor/django-compressor
# django-redis>=4.11.0  # https://github.com/niwinz/django-redis
# Django REST Framework
djangorestframework>=3.11.0,<3.12.0  # https://github.com/encode/django-rest-framework

django-background-tasks
sentry-sdk>=0.14.3,<0.15.0
django-cors-headers>=3.2.1,<4.0.0
django-partial-date>=1.3.0,<2.0.0

django-tables2>=2.3.1,<3.0.0
django-simple-history>=2.8.0,<3.0.0
django-import-export>=2.0.2,<3.0.0
# django-autocomplete-light>=3.5.1,<4.0.0
 git+https://github.com/nad2000/django-autocomplete-light.git
xhtml2pdf>=0.2.4,<1.0.0
django-select2>=7.2.3,<8.0.0
django-extra-views>=0.13.0,<1.0.0
django-private-storage>=2.2.2,<3.0.0
django-fsm>=2.7.0,<3.0.0
django-fsm-admin>=1.2.4,<2.0.0
django-fsm-log>=2.0.1,<3.0.0
graphviz>=0.4

django-summernote>=0.8.11.6,<0.9.0.0

django-modeltranslation==0.15
django-recaptcha==2.0.6
jinja2>=2.11.2,<2.12.0
django-weasyprint==1.0.1
html2text==2020.1.16
